feat(contractStorage): introduce rocksdb additionally to sqlite and make it default
