+++ crates/locutus-core/Cargo.toml
@@ -46,7 +46,9 @@ tokio = { version = "1", features = ["rt-multi-thread", "sync", "macros", "fs"] 
 unsigned-varint = "0.7"
 uuid = { version = "1", features = ["serde", "v4", "v1"] }
 rmp-serde = { workspace = true, optional = true }
-sqlx = { version = "0.6", features = ["sqlite", "runtime-tokio-rustls"] }
+sqlx = { version = "0.6", features = ["sqlite", "runtime-tokio-rustls"], optional = true }
+# TODO(kakoc): clang should be installed for rocksdb; write about that in prerequisites/dev guide
+rocksdb = { version = "0.19.0", default-features = false, optional = true }
 warp = { version = "0.3", default-features = false, optional = true, features = ["multipart", "tls"] }
 
 # Tracing deps
@@ -70,6 +72,8 @@ locutus-stdlib = { path = "../locutus-stdlib", features = ["testing", "net"] }
 
 [features]
 testing = ["arbitrary"]
-default = ["websocket", "trace"]
+default = ["websocket", "rocks_db", "trace"]
+rocks_db = ["rocksdb"]
+sqlite = ["sqlx"]
 websocket = ["warp/websocket", "rmp-serde"]
 trace = ["tracing", "opentelemetry", "opentelemetry-jaeger", "tracing-opentelemetry", "tracing-subscriber"]
