+++ crates/locutus-node/Cargo.toml
@@ -35,6 +35,6 @@ rmpv = { workspace = true }
 rmp-serde = { workspace = true }
 
 # internal:
-locutus-core = { path = "../locutus-core", features = ["websocket"], version = "0.0.3" }
+locutus-core = { path = "../locutus-core", version = "0.0.3" }
 locutus-dev = { path = "../locutus-dev", version = "0.0.3" }
 locutus-stdlib = { path = "../locutus-stdlib", version = "0.0.3" }
