+++ crates/locutus-node/src/bin/node.rs
@@ -1,7 +1,7 @@
 use clap::Parser;
 use locutus_core::{
     locutus_runtime::{ContractStore, StateStore},
-    Config, Executor, SqlitePool,
+    Config, Executor, Storage,
 };
 use locutus_dev::config::OperationMode;
 use std::net::SocketAddr;
@@ -28,7 +28,7 @@ async fn run_local(config: NodeConfig) -> Result<(), DynError> {
         .unwrap_or_else(|| Config::get_conf().config_paths.local_contracts_dir());
     // TODO: Generate components and secrets store from config
     let contract_store = ContractStore::new(contract_dir, MAX_SIZE)?;
-    let state_store = StateStore::new(SqlitePool::new().await?, MAX_MEM_CACHE).unwrap();
+    let state_store = StateStore::new(Storage::new().await?, MAX_MEM_CACHE).unwrap();
     let executor = Executor::new(contract_store, state_store, || {
         locutus_core::util::set_cleanup_on_exit().unwrap();
     })
